requirements.txt
streamlit
spotipy
python-dotenv
pandas
